Restructure and add argument parser
Split up script into constituent components.
This will make it easier to add features.

- Add option to change projection #1
